chore(toast-utils): move file to ui.utils folder
