+++ app/src/main/java/com/android/unio/ui/association/EditAssociation.kt
@@ -43,9 +43,9 @@ import com.android.unio.model.association.Association
 import com.android.unio.model.association.AssociationCategory
 import com.android.unio.model.association.AssociationViewModel
 import com.android.unio.model.strings.test_tags.EditAssociationTestTags
-import com.android.unio.model.utils.ToastUtils
 import com.android.unio.ui.navigation.NavigationAction
 import com.android.unio.ui.navigation.Screen
+import com.android.unio.ui.utils.ToastUtils
 import kotlinx.coroutines.launch
 
 @Composable
